backend/requirements.txt
# Telegram Bot
aiogram>=3.24.0

# Job Queue
rq>=1.16.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# TikTok Scraping
yt-dlp>=2024.1.0
httpx>=0.27.0

# LLM
google-generativeai>=0.8.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
